Bug fixed in tests
