+++ tests/Product/ProductTest.php
@@ -9,7 +9,7 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 
-class ProductTestCase extends ModelTestCase
+class ProductTest extends ModelTestCase
 {
     protected string $model = Product::class;
     protected string $factory = ProductFactory::class;
